Rename user_events.py to devices.py
